refactor: 구조 개선
