+++ .storybook/preview.tsx
@@ -1,11 +1,9 @@
-import type { Preview } from "@storybook/react";
-import { ThemeProvider } from "styled-components";
-import theme from "../src/styles/theme";
-import { GlobalStyle } from "../src/styles/GlobalStyle";
+import type {Preview} from "@storybook/react";
+import {GlobalStyle} from "../src/styles/GlobalStyle";
 
 const preview: Preview = {
   parameters: {
-    actions: { argTypesRegex: "^on[A-Z].*" },
+    actions: {argTypesRegex: "^on[A-Z].*"},
     controls: {
       matchers: {
         color: /(background|color)$/i,
@@ -15,10 +13,10 @@ const preview: Preview = {
   },
   decorators: [
     (Story) => (
-      <ThemeProvider theme={theme}>
-        <GlobalStyle />
-        <Story />
-      </ThemeProvider>
+      <>
+        <GlobalStyle/>
+        <Story/>
+      </>
     ),
   ],
 };
