--- lib/index.ts
@@ -1,3 +1,0 @@
-export {default as Container} from "../src/layouts/Container";
-export {default as Flex} from "../src/layouts/Flex";
-export {default as Grid} from "../src/layouts/Grid";
