--- src/layouts/Container/index.ts
@@ -1,3 +1,0 @@
-import Container from "./Container";
-
-export default Container;
