--- src/layouts/Flex/index.ts
@@ -1,2 +1,0 @@
-import Flex from "./Flex";
-export default Flex;
