--- src/layouts/Grid/index.ts
@@ -1,3 +1,0 @@
-import Grid from "./Grid";
-
-export default Grid;
