+++ src/lib/layouts/Flex/index.ts
@@ -1,0 +1,3 @@
+import Flex from "./Flex.tsx";
+
+export default Flex;
