+++ src/lib/layouts/Grid/index.ts
@@ -1,0 +1,3 @@
+import Grid from "./Grid.tsx";
+
+export default Grid;
