+++ src/main.tsx
@@ -1,16 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { GlobalStyle } from "./styles/GlobalStyle.ts";
+import {GlobalStyle} from "./styles/GlobalStyle.ts";
 import router from "./router.tsx";
-import { RouterProvider } from "react-router-dom";
-import { ThemeProvider } from "styled-components";
-import theme from "./styles/theme.ts";
+import {RouterProvider} from "react-router-dom";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <ThemeProvider theme={theme}>
-      <GlobalStyle />
-      <RouterProvider router={router} />
-    </ThemeProvider>
+    <GlobalStyle/>
+    <RouterProvider router={router}/>
   </React.StrictMode>
 );
