--- src/styles/theme.ts
@@ -1,4 +1,0 @@
-const theme = {
-  spacingUnit: 8,
-};
-export default theme;
