+++ tsconfig.json
@@ -25,8 +25,7 @@
     "noFallthroughCasesInSwitch": true
   },
   "include": [
-    "src",
-    "src/lib"
+    "src"
   ],
   "references": [
     {
